+++ elements/pl-big-o-input/pl-big-o-input.py
@@ -33,7 +33,7 @@ SIZE_DEFAULT = 35
 PLACEHOLDER_TEXT_THRESHOLD = 20
 SHOW_HELP_TEXT_DEFAULT = True
 WEIGHT_DEFAULT = 1
-DISPLAY_DEFAULT = bou.DisplayType.INLINE
+DISPLAY_DEFAULT = pl.DisplayType.INLINE
 BIG_O_TYPE_DEFAULT = BigOType.BIG_O
 
 
@@ -81,10 +81,10 @@ def render(element_html: str, data: pl.QuestionData) -> str:
     variables = phs.get_variables_list(
         pl.get_string_attrib(element, "variable", VARIABLES_DEFAULT)
     )
-    display = bou.get_enum_attrib(element, "display", bou.DisplayType, DISPLAY_DEFAULT)
+    display = pl.get_enum_attrib(element, "display", pl.DisplayType, DISPLAY_DEFAULT)
     size = pl.get_integer_attrib(element, "size", SIZE_DEFAULT)
 
-    bigo_type = bou.get_enum_attrib(element, "type", BigOType, BIG_O_TYPE_DEFAULT).value
+    bigo_type = pl.get_enum_attrib(element, "type", BigOType, BIG_O_TYPE_DEFAULT).value
 
     constants_class = phs._Constants()
 
@@ -130,7 +130,7 @@ def render(element_html: str, data: pl.QuestionData) -> str:
         }
 
         if score is not None:
-            score_type, score_value = bou.determine_score_params(score)
+            score_type, score_value = pl.determine_score_params(score)
             html_params[score_type] = score_value
 
         with open("pl-big-o-input.mustache", "r", encoding="utf-8") as f:
@@ -181,7 +181,7 @@ def render(element_html: str, data: pl.QuestionData) -> str:
         }
 
         if score is not None:
-            score_type, score_value = bou.determine_score_params(score)
+            score_type, score_value = pl.determine_score_params(score)
             html_params[score_type] = score_value
 
         with open("pl-big-o-input.mustache", "r", encoding="utf-8") as f:
@@ -245,9 +245,9 @@ def grade(element_html: str, data: pl.QuestionData) -> None:
     if a_tru is None:
         return
 
-    big_o_type = bou.get_enum_attrib(element, "type", BigOType, BIG_O_TYPE_DEFAULT)
+    big_o_type = pl.get_enum_attrib(element, "type", BigOType, BIG_O_TYPE_DEFAULT)
 
-    bou.grade_answer_parameterized(
+    pl.grade_answer_parameterized(
         data,
         name,
         lambda a_sub: GRADE_FUNCTION_DICT[big_o_type](a_tru, a_sub, variables),
@@ -274,7 +274,7 @@ def test(element_html: str, data: pl.ElementTestData) -> None:
 
     elif result == "incorrect":
         data["raw_submitted_answers"][name] = f"{random.randint(4, 100):d} * {a_tru}"
-        bigo_type = bou.get_enum_attrib(element, "type", BigOType, BIG_O_TYPE_DEFAULT)
+        bigo_type = pl.get_enum_attrib(element, "type", BigOType, BIG_O_TYPE_DEFAULT)
 
         if bigo_type is BigOType.THETA:
             data["partial_scores"][name] = {
