+++ python/python_helper_sympy_test.py
@@ -1,8 +1,9 @@
 import json
+from typing import List
+
 import pytest
 import python_helper_sympy as phs
 import sympy
-from typing import List
 
 
 class TestSympy:
@@ -33,7 +34,7 @@ class TestSympy:
         "ln(m)",
         "atanh(m)",
         "asinh(m)",
-        "acosh(m)"
+        "acosh(m)",
     ]
 
     EXPR_LIST: List = [
@@ -61,7 +62,7 @@ class TestSympy:
         sympy.log(M),
         sympy.atanh(M),
         sympy.asinh(M),
-        sympy.acosh(M)
+        sympy.acosh(M),
     ]
 
     @pytest.mark.parametrize("a_sub, sympy_ref", zip(EXPR_STRINGS, EXPR_LIST))
